working on CreditCardRequest to get it to use the new validation.

+++ PelotonEppSdk/Models/CreditCardRequest.cs
@@ -13,7 +13,7 @@ using static PelotonEppSdk.Validations.RequestMethodAttribute;
 
 namespace PelotonEppSdk.Models
 {
-    public class CreditCardRequest : RequestBase, ICreditCardCreateRequest, ICreditCardDeleteRequest, ICreditCardUpdateRequest
+    public class CreditCardRequest : RequestBase//, ICreditCardCreateRequest, ICreditCardDeleteRequest, ICreditCardUpdateRequest
     {
         public CreditCardRequest()
         {
@@ -21,6 +21,8 @@ namespace PelotonEppSdk.Models
         }
 
         [RequestMethod(new[] { RequestMethodEnum.GET, RequestMethodEnum.PUT, RequestMethodEnum.DELETE })]
+        [Required]
+        [StringLength(32, MinimumLength = 32)]
         public string CreditCardToken { get; set; }
 
         /// <summary>
@@ -118,14 +120,22 @@ namespace PelotonEppSdk.Models
         }
 
         /// <exception cref="HttpException"><see cref="HttpStatusCode"/> is not <c>2XX Success</c>.</exception>
-        public async Task<CreditCardResponse> PostAsync()
-        {
-            return await PostAsync(false).ConfigureAwait(false);
-        }
+//        public async Task<CreditCardResponse> PostAsync()
+//        {
+//            return await PostAsync(false).ConfigureAwait(false);
+//        }
 
         /// <exception cref="HttpException"><see cref="HttpStatusCode"/> is not <c>2XX Success</c>.</exception>
-        public async Task<CreditCardResponse> PutAsync()
+        public async Task<CreditCardResponse> PutAsync(bool validate = true)
         {
+            if (validate)
+            {
+                // validate the request before sending it
+                ValidationErrors = new List<string>();
+                var isValid = TryValidatePropertySubset(ValidationErrors, RequestMethodEnum.POST);
+                if(!isValid) throw new ValidationException("Validation Error", null, ValidationErrors);
+            }
+
             var client = new PelotonClient();
             var request = (credit_card_request) this;
             var result = await client.PutAsync<credit_card_response>(request, ApiTarget.CreditCards, CreditCardToken).ConfigureAwait(false);
@@ -133,8 +143,16 @@ namespace PelotonEppSdk.Models
         }
 
         /// <exception cref="HttpException"><see cref="HttpStatusCode"/> is not <c>2XX Success</c>.</exception>
-        public async Task<CreditCardResponse> DeleteAsync()
+        public async Task<CreditCardResponse> DeleteAsync(bool validate = true)
         {
+            if (validate)
+            {
+                // validate the request before sending it
+                ValidationErrors = new List<string>();
+                var isValid = TryValidatePropertySubset(ValidationErrors, RequestMethodEnum.POST);
+                if(!isValid) throw new ValidationException("Validation Error", null, ValidationErrors);
+            }
+
             var client = new PelotonClient();
             var request = (credit_card_request) this;
             var result = await client.DeleteAsync<credit_card_response>(request, ApiTarget.CreditCards, CreditCardToken).ConfigureAwait(false);
